+++ src/rotp/model/ships/ShipWeaponBiological.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2015-2020 Ray Fowler
- * Modifications Copyright 2023-2024 Ilya Zushinskiy
+ * Modifications Copyright 2023-2025 Ilya Zushinskiy
  * 
  * Licensed under the GNU General Public License, Version 3 (the "License");
  * you may not use this file except in compliance with the License.
@@ -57,8 +57,8 @@ public final class ShipWeaponBiological extends ShipWeapon {
     @Override
     public void fireUpon(CombatEntity source, CombatEntity target, int count) {
         float defense = target.bioweaponDefense();
-        float attack = source.attackLevel();
-        float pct = (5 + attack - defense) / 10;
+        int attack = source.attackLevel();
+        float pct = (5 + attack - defense) / 10f;
         pct = max(.05f, pct);
 
         float totalDamage = 0;
