+++ src/rotp/model/ships/ShipWeaponTorpedo.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2015-2020 Ray Fowler
- * Modifications Copyright 2023 Ilya Zushinskiy
+ * Modifications Copyright 2023-2025 Ilya Zushinskiy
  * 
  * Licensed under the GNU General Public License, Version 3 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ public final class ShipWeaponTorpedo extends ShipWeaponMissileType {
     @Override
     public TechTorpedoWeapon tech()       { return (TechTorpedoWeapon) super.tech(); }
     @Override
-    public float computerLevel()         { return tech().computer; }
+    public int computerLevel()            { return tech().computer; }
     @Override
     public float damageLoss(float dist) { return dist*tech().damageLoss; }
     @Override
